+++ RELEASE-NOTES.md
@@ -1,5 +1,14 @@
 # Release Notes
 
+## v1.0.6 - 2018-12-4
+Improve error handling
+Publish DID in public profile
+Doc fixes and cleanup
+
+Bug fixes:
+* running getProfile in node
+* libp2p bootstrap config errors
+
 ## v1.0.5 - 2018-11-14
 Added example for getProfile
 Some bug fixes: